+++ spec/integrations/pro/performance_tracking_spec.rb
@@ -42,7 +42,8 @@ start_karafka_and_wait_until do
 end
 
 TOPICS.each do |topic_name|
-  p95 = Karafka::Pro::PerformanceTracker.instance.processing_time_p95(topic_name, 0)
+  tracker = Karafka::Pro::Instrumentation::PerformanceTracker.instance
+  p95 = tracker.processing_time_p95(topic_name, 0)
 
   message_speed = MESSAGE_SPEED.fetch(topic_name)