+++ spec/lib/karafka/instrumentation/vendors/datadog/logger_listener_spec.rb
@@ -1,4 +1,12 @@
 # frozen_string_literal: true
 
-# Placeholder for Coditsu
-# This is covered in the integration suite
+require 'karafka/instrumentation/vendors/datadog/logger_listener'
+
+# This is fully covered in the integration suite
+RSpec.describe_current do
+  subject(:listener) { described_class.new }
+
+  describe 'events mapping' do
+    it { expect(NotificationsChecker.valid?(listener)).to eq(true) }
+  end
+end